MyRect 구조체의 생성자 매개변수 origin을 원점이 아닌 기준점으로 변경
기준점은 iOS Frame과 동일하게 leftTop으로 설정.

+++ CoordinateCalculator/Sources/CoordinateCalculator/MyPoint.swift
@@ -9,15 +9,15 @@
 import Foundation
 
 struct MyPoint {
-    private(set) var x: Float
-    private(set) var y: Float
+    private(set) var x: Int
+    private(set) var y: Int
     
     init() {
         self.x = 0
         self.y = 0
     }
     
-    init(x: Float, y: Float) {
+    init(x: Int, y: Int) {
         self.x = x
         self.y = y
     }
